+++ src/components/Event.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace mix8872\config\components;
+
+
+class Event extends \yii\base\Event
+{
+    public $model;
+}