+++ src/views/default/create.php
@@ -4,4 +4,4 @@ $this->title = Yii::t('config', 'Добавление элемента');
 $this->params['breadcrumbs'][] = ['label' => 'Настройки', 'url' => ['index']];
 $this->params['breadcrumbs'][] = $this->title;
 
-echo $this->render('_form', compact('model', 'groups', 'tabs', 'tabModel'));
+echo $this->render('_form', compact('model', 'groups', 'tabs', 'tabModel', 'authItems'));